remove share button from each category
